requirements.txt
beautifulsoup4==4.13.3
mysql_connector_repackaged==0.3.1
prettytable==3.16.0
PyMySQL==1.1.1
Requests==2.32.3
